Break material state module cycle
